tools/ani_gen_tools/sprite_processor/requirements.txt
Pillow>=8.0.0
PyQt5>=5.15.0
